Move camera logic to view module, adjust follow parameters to account for smaller viewport
