+++ src/player/plugin.rs
@@ -48,13 +48,8 @@ impl Plugin for PlayerPlugin {
                         player_movement,
                         update_player_aim_position,
                         on_player_experience_change,
-                        camera_debug_system,
                     )
                         .in_set(InGameSet::Simulation),
-                    // avian recommended ordering for camera following logic
-                    camera_follow_system
-                        .in_set(InGameSet::Camera)
-                        .before(TransformSystem::TransformPropagate),
                     (draw_cursor, animate_level_up).in_set(InGameSet::Vfx),
                 ),
             )
