+++ src/player/systems/mod.rs
@@ -1,4 +1,3 @@
-pub mod camera_follow;
 pub mod consume;
 pub mod cursor;
 pub mod death;
@@ -7,7 +6,6 @@ pub mod input;
 pub mod movement;
 pub mod spawn_player;
 
-pub use camera_follow::*;
 pub use consume::*;
 pub use cursor::*;
 pub use handle_level_up::*;
